fix(auth): add dynamic export to resolve useSearchParams build errors
